Move colors to params
